refactor(deploy): give the result its own module and one copy affordance
The verified-success card grew the Timeline pane past 1.8k lines by
reaching into helpers only it used, and copied the task-id row's
click-to-copy timer a second time. Both halves move out: the result card
lives in deployment-task-success-section.tsx, and the clipboard
feedback — a check mark that appears only once the browser accepted the
write — becomes one hook with one timer that both rows share.

Drops `autoCloseOnSuccess` on the way out: no host ever passed it, and
"the Timeline gets out of the way after it delivered the result" is the
behavior, not an option.
